[dns] fix name length check in `Name::AppendMultipleLabels()`
Corrects an off-by-one error in the length validation within the
`Name::AppendMultipleLabels()` method.

The previous check (`index + 1 < kMaxEncodedLength`) would incorrectly
return an `kErrorInvalidArgs` error if the encoded name length was
exactly equal to `kMaxEncodedLength`.

The check is updated to `index + 1 <= kMaxEncodedLength` to allow for
names of the maximum valid length. This issue was discovered by fuzz
testing.