[tcat] Fix TCAT Commissioner authorization checks; add unit test
In previous code, only one of the Active Dataset or newly provided dataset by the Commissioner
were selected to do authorization checks with. This is now fixed by considering both of
these datasets, if available. If an item to check like XPAN ID or Network Name is absent
from the Active Dataset, it will now be taken from the Commissioner-provided new
dataset (called the 'Backup Dataset' for the purpose of the checks), if present.
If the item is absent from both datasets then the check will be considered successful
if at least the TCAT Commissioner's certificate contains the item.

A unit test is added for checking Commissioner authorization. To do these checks, a new test
class TcatAgentTestProbe is added so that all relevant conditions can be easily tested.