[tcat] Update TCAT Commissioner authorization checks; add unit tests
Updates the TCAT class public methods for doing Commissioner authorization checks
and clarifies the code, with minor updates to PSKc cases handling.

Unit tests are added for checking Commissioner authorization. To do these checks, a new test
class TcatAgentTestProbe is added so that all relevant conditions can be easily tested.

Also reverts the CommCert4 fix that was made in #12151.