[tcat] Update TCAT Commissioner authorization checks; add unit tests (#12182)
Updates the TCAT class public methods for doing Commissioner authorization checks
and clarifies the code, with minor updates to PSKc cases handling.

Unit tests are added for checking Commissioner authorization. To do these checks, a new test
class UnitTester is added which has access to private members of the TcatAgent class.
Validation/mock functions are added in the test code to keep the unit tests readable.

Also reverts the CommCert4 fix that was made in #12151.

For more background information see JIRA BHC-766.